+++ crates/webauthn/tests/src/auth/mod.rs
@@ -1,15 +1,10 @@
-use cairo_args_runner::{arg_array, arg_value, felt_vec, Arg, Felt252};
+use crate::prelude::*;
 
 mod expand_auth_data;
 mod helpers;
 mod mod_arithmetic;
 mod verify_ecdsa;
 mod verify_signature;
-
-#[derive(Debug)]
-pub struct ArgsBuilder {
-    args: Vec<Arg>,
-}
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct CairoU256 {
@@ -67,34 +62,6 @@ impl FeltSerialize for P256r1PublicKey {
             .chain(self.y.to_felts())
             .collect()
     }
-}
-
-impl ArgsBuilder {
-    pub fn new() -> Self {
-        Self { args: Vec::new() }
-    }
-    #[allow(dead_code)]
-    pub fn add_one(mut self, arg: impl Into<Felt252>) -> Self {
-        self.args.push(Arg::Value(arg.into()));
-        self
-    }
-    pub fn add_struct(mut self, args: impl IntoIterator<Item = impl Into<Felt252>>) -> Self {
-        self.args
-            .extend(args.into_iter().map(|arg| Arg::Value(arg.into())));
-        self
-    }
-    pub fn add_array(mut self, args: impl IntoIterator<Item = impl Into<Felt252>>) -> Self {
-        self.args
-            .push(Arg::Array(args.into_iter().map(|arg| arg.into()).collect()));
-        self
-    }
-    pub fn build(self) -> Vec<Arg> {
-        self.args
-    }
-}
-
-pub trait FeltSerialize {
-    fn to_felts(self) -> Vec<Felt252>;
 }
 
 #[test]
