+++ crates/webauthn/tests/src/auth/mod_arithmetic.rs
@@ -1,7 +1,6 @@
 use cairo_args_runner::{Arg, Felt252};
 
-use crate::*;
-use crate::{ConstLenExtractor, Function};
+use crate::prelude::*;
 
 /// ```extended_gcd(u256, u256) -> (u256, u256, u256)```
 const EXTENDED_GCD: Function<GCDParser, ConstLenExtractor<6>> =
