+++ crates/webauthn/tests/src/lib.rs
@@ -1,7 +1,8 @@
 #[cfg(test)]
+mod arg_builder;
+#[cfg(test)]
 mod auth;
-
+#[cfg(test)]
+pub mod prelude;
 #[cfg(test)]
 mod utils;
-#[cfg(test)]
-pub use utils::*;
